// src/app/sistema/bienestar/interfaces/fichaVivienda.ts
export interface FichaVivienda {
  iViendaCarId: number;
  iFichaDGId: number;
  iTipoOcupaVivId: number;
  iMatPreId: number;
  iTipoVivId: number;
  iViviendaCarNroPisos: number;
  iViviendaCarNroAmbientes: number;
  iViviendaCarNroHabitaciones: number;
  iEstadoVivId: number;
  iMatPisoVivId: number;
  iMatTecVivId: number;
  iTiposSsHhId: number;
  iTipoSumAId: number;
}